requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
langchain>=0.3.0
langchain-google-vertexai>=2.0.0
langchain-core>=0.3.0
langgraph>=0.2.0
pandas>=2.0.0
plotly>=5.17.0
matplotlib>=3.7.0
numpy>=1.26.0
python-multipart>=0.0.6
redis>=5.0.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
docker>=6.0.0
websockets>=12.0
aiofiles>=23.0.0
jinja2>=3.1.0
setuptools>=68.0.0